feat: #80 節約時の残高更新にアニメーションを付与
節約時の残高更新にアニメーションが無く、ためている感がなかったため、アニメーションを付与した
