+++ src/app/_components/Loading.tsx
@@ -7,7 +7,7 @@ export const Loading = () => {
       <div
         role="spinner"
         aria-label="スピナー"
-        className="h-16 w-16 animate-spin rounded-full border-4 border-pink-500 border-t-transparent sm:h-20 sm:w-20"
+        className="h-16 w-16 animate-spin rounded-full border-4 border-green-400 border-t-transparent sm:h-20 sm:w-20"
       ></div>
     </div>
   );
