+++ src/app/main/_components/_BalanceDisplayAndAdd/BalanceContainer.tsx
@@ -1,20 +1,13 @@
 import React from 'react';
 
 import { getBalance, getSessionAndUserId } from '@/app/lib/commonFunction';
+import { BalanceDisplayAndAdd } from './BalanceDisplayAndAdd';
 
-export const Balance = async () => {
+export const BalanceContainer = async () => {
   try {
-    // UserID取得
     const userId = await getSessionAndUserId();
-    // 残高取得
     const balance = await getBalance(userId);
-
-    // 取得した残高を表示
-    return (
-      <h1 aria-label="balance" className="text-gray-700">
-        {balance.balance.toLocaleString()}
-      </h1>
-    );
+    return <BalanceDisplayAndAdd balance={balance.balance} />;
   } catch {
     throw new Error('残高の取得中にエラーが発生しました。');
   }
